compilation

+++ gurobi_c.c
@@ -20,7 +20,7 @@ int vcrToidx(int v, int c, int r, int dim){
 	return idx;
 }
 
-int get_gurobi_solution(int *sol, int block_h, int block_w)
+int get_gurobi_solution(double *sol, int block_h, int block_w)
 {
   GRBenv   *env   = NULL;
   GRBmodel *model = NULL;
